Add CSS file and style to all recipe subsites
In this commit I added style.css file with complete styling for
main site, index.html as well as for the subsites for each recipe.
Styling focuses on background color and fonts weight and colors.
I also modified list on the main site - changing it to buttons so it
looks a little bit better.